Split the User resource & JWT interfaces
We still extend the User interface so that we don't
break anyone's build, but added a TODO to properly
break the link in the next majar.

Change-type: patch
